Add business logic

+++ src/main/java/my/anatoliy57/gasstation/domain/dto/PeriodDto.java
@@ -13,6 +13,6 @@ public class PeriodDto {
     private Long id;
     private Long stationId;
     private Integer density;
-    private Long startTime;
-    private Long endTime;
+    private Integer startTime;
+    private Integer endTime;
 }
